requirements.txt
feedparser
requests
pyyaml
discord.py
lxml